[DEV-2268] remove btc/usdt amount in cryptoInput (#1165)
* [DEV-2268] remove btc/usdt amount in cryptoInput

* [DEV-2268] add migration

* [DEV-2268] Refactoring

* [DEV-2268] remove more

---------
