+++ src/subdomains/supporting/payin/factories/payin.factory.ts
@@ -5,8 +5,8 @@ import { PayInEntry } from '../interfaces';
 @Injectable()
 export class PayInFactory {
   createFromEntry(entry: PayInEntry): CryptoInput {
-    const { address, txId, txType, txSequence, blockHeight, amount, asset, btcAmount, usdtAmount } = entry;
+    const { address, txId, txType, txSequence, blockHeight, amount, asset } = entry;
 
-    return CryptoInput.create(address, txId, txType, txSequence, blockHeight, amount, asset, btcAmount, usdtAmount);
+    return CryptoInput.create(address, txId, txType, txSequence, blockHeight, amount, asset);
   }
 }
