+++ src/subdomains/supporting/payin/strategies/register/impl/optimism.strategy.ts
@@ -2,12 +2,8 @@ import { Injectable, OnModuleInit } from '@nestjs/common';
 import { Config } from 'src/config/config';
 import { AlchemyNetworkMapper } from 'src/integration/alchemy/alchemy-network-mapper';
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
-import { Asset, AssetType } from 'src/shared/models/asset/asset.entity';
-import { AssetService } from 'src/shared/models/asset/asset.service';
-import { RepositoryFactory } from 'src/shared/repositories/repository.factory';
 import { DfxLogger } from 'src/shared/services/dfx-logger';
 import { QueueHandler } from 'src/shared/utils/queue-handler';
-import { PayInRepository } from '../../../repositories/payin.repository';
 import { PayInOptimismService } from '../../../services/payin-optimism.service';
 import { EvmStrategy } from './base/evm.strategy';
 
@@ -15,13 +11,8 @@ import { EvmStrategy } from './base/evm.strategy';
 export class OptimismStrategy extends EvmStrategy implements OnModuleInit {
   protected readonly logger = new DfxLogger(OptimismStrategy);
 
-  constructor(
-    optimismService: PayInOptimismService,
-    payInRepository: PayInRepository,
-    assetService: AssetService,
-    repos: RepositoryFactory,
-  ) {
-    super('ETH', optimismService, payInRepository, assetService, repos);
+  constructor(optimismService: PayInOptimismService) {
+    super(optimismService);
   }
 
   onModuleInit() {
@@ -41,31 +32,6 @@ export class OptimismStrategy extends EvmStrategy implements OnModuleInit {
 
   get blockchain(): Blockchain {
     return Blockchain.OPTIMISM;
-  }
-
-  //*** PUBLIC API ***//
-
-  async getReferenceAssets(): Promise<{ btc: Asset; usdt: Asset }> {
-    return Promise.all([
-      this.assetService.getAssetByQuery({
-        dexName: 'WBTC',
-        blockchain: Blockchain.ETHEREUM,
-        type: AssetType.TOKEN,
-      }),
-      this.assetService.getAssetByQuery({
-        dexName: 'USDT',
-        blockchain: Blockchain.ETHEREUM,
-        type: AssetType.TOKEN,
-      }),
-    ]).then(([btc, usdt]) => ({ btc, usdt }));
-  }
-
-  async getSourceAssetRepresentation(asset: Asset): Promise<Asset> {
-    return this.assetService.getAssetByQuery({
-      dexName: asset.dexName,
-      blockchain: Blockchain.ETHEREUM,
-      type: asset.type,
-    });
   }
 
   //*** HELPER METHODS ***//
